+++ frontend/src/store/SpecialVersion.ts
@@ -1,0 +1,3 @@
+// 由 Nova 自行维护的 MC 特供版本～
+export const Special_MC_Version = [
+]